Time: 4 ms (22.22%), Space: 6 MB (27.65%) - LeetHub

+++ 0693-binary-number-with-alternating-bits/0693-binary-number-with-alternating-bits.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     bool hasAlternatingBits(int n) 
     {
-        long  k=n;
+        long long k=n;
         k>>=1;
         
         //cout<<k+n<<endl;
-        long newnum=k+n;
+        long long newnum=k+n;
        return ((newnum+1)&newnum)==0?1:0;
     }
 };
